+++ waste/memory.go
@@ -14,7 +14,7 @@ type GiBObject struct {
 	B [GiB]byte
 }
 
-func Memory(gib int) {
+func Memory(gib float64) {
 	Buffers = make([]*GiBObject, 0, gib)
 	for gib > 0 {
 		o := new(GiBObject)
